requirements.txt
asgiref==3.5.2
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
cloudinary==1.30.0
cryptography==38.0.4
defusedxml==0.7.1
Django==4.0.4
django-admin-interface==0.23.0
django-ckeditor==6.5.1
django-cloudinary-storage==0.3.0
django-colorfield==0.8.0
django-cors-headers==3.13.0
django-crispy-forms==1.14.0
django-flat-responsive==2.0
django-flat-theme==1.1.4
django-js-asset==2.0.0
djangorestframework==3.14.0
gunicorn==20.1.0
idna==3.4
oauthlib==3.2.2
Pillow==9.3.0
psycopg2-binary==2.9.5
pycparser==2.21
PyJWT==2.6.0
python-dotenv==0.21.0
python3-openid==3.2.0
pytz==2022.6
requests==2.28.1
requests-oauthlib==1.3.1
six==1.16.0
sqlparse==0.4.3
tzdata==2022.6
urllib3==1.26.13
whitenoise==6.3.0
